+++ redis/server.hh
@@ -91,12 +91,8 @@ private:
     class connection : public generic_server::connection {
         redis_server& _server;
         socket_address _server_addr;
-        input_stream<char> _read_buf;
-        output_stream<char> _write_buf;
         redis_protocol_parser _parser;
-        seastar::gate _pending_requests_gate;
         redis::redis_options _options;
-        future<> _ready_to_respond = make_ready_future<>();
 
         using execution_stage_type = inheriting_concrete_execution_stage<
                 future<redis_server::result>,
@@ -109,9 +105,8 @@ private:
     public:
         connection(redis_server& server, socket_address server_addr, connected_socket&& fd, socket_address addr);
         virtual ~connection();
-        future<> process();
-        future<> process_request();
-        void handle_error(future<>&& f);
+        future<> process_request() override;
+        void handle_error(future<>&& f) override;
         void write_reply(const redis_exception&);
         void write_reply(redis_server::result result);
     private:
